requirements.txt
setuptools>=65.5.0
wheel>=0.38.4
flask>=2.0.1
python-dotenv>=0.19.0
neo4j>=5.11.0
langchain>=0.0.200
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=0.25.0
requests>=2.31.0
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
jieba>=0.42.1
gunicorn>=21.2.0
pydantic>=1.10.13
